Server-side support for multi-browser configurations.  If multiple browsers (comma-separated) are defined in the "browser" definition for a given location then they will be exploded into virtual locations that share a work queue and the job sent to the test machine will include a browser=X line telling wptdriver which browser to use.  This means that the browser sections in wptdriver.ini will have to match the browsers listed in the locations.ini file and that all machines sharing a given "location" will support the same list of browsers but I think those are reasonable requirements.
This is also backwards compatible in that existing locations with just a single browser will not be treated any differently.  Initial testing has worked but I want to do some more robust testing before pushing it live (and then it should be easy to add the support on the agent side to load the browser settings as each test comes in instead of at launch time).